requirements.txt
langchain
langchain-groq
langchain-experimental
langchain-community
pandas
tabulate
python-dotenv
streamlit